backend/pyproject.toml
[project]
name = "docsense"
version = "0.1.0"
description = "Production-grade document intelligence platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "bcrypt>=5.0.0",
    "pymupdf>=1.27.1",
    "sentence-transformers>=5.2.3",
    "qdrant-client>=1.17.0",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
eval = [
    "ragas>=0.4.3",
    "datasets>=4.6.0",
    "langchain-huggingface>=1.2.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
